Display invitation status on console
This commit adds functionality to display the invitation status being sent to
the individual from the search results. This change makes the interface more
engaging to the user as the user can quantify the acheived invitees by looking
at the console.

Tests had been added to ensure the reliability of the "invitation" module.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>